+++ app/Services/Contact/Contact/UpdateWorkInformation.php
@@ -37,6 +37,7 @@ class UpdateWorkInformation extends BaseService
     {
         $this->validate($data);
 
+        /** @var Contact */
         $contact = Contact::where('account_id', $data['account_id'])
             ->findOrFail($data['contact_id']);