+++ app/Services/Contact/Conversation/UpdateConversation.php
@@ -38,6 +38,7 @@ class UpdateConversation extends BaseService
     {
         $this->validate($data);
 
+        /** @var Conversation */
         $conversation = Conversation::where('account_id', $data['account_id'])
                                     ->findOrFail($data['conversation_id']);