+++ resources/lang/el/changelog.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * ⚠️ Edition not allowed except for 'en' language.
+ *
+ * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
+ */
+
+return [
+    'title' => 'Αλλαγές προϊόντος',
+    'note' => 'Σημείωση: Δυστυχώς αυτή η σελίδα είναι διαθέσιμη μόνο στα Αγγλικά.',
+];