+++ resources/lang/vi/changelog.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * ⚠️ Edition not allowed except for 'en' language.
+ *
+ * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
+ */
+
+return [
+    'title' => 'Product changes',
+    'note' => 'Note: unfortunately, this page is only in English.',
+];