+++ resources/lang/vi/dashboard.php
@@ -1,0 +1,42 @@
+<?php
+
+/**
+ * ⚠️ Edition not allowed except for 'en' language.
+ *
+ * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
+ */
+
+return [
+    'dashboard_blank_title' => 'Chào mừng đến với tài khoản của bạn!',
+    'dashboard_blank_description' => 'Monica là nơi sắp xếp tất cả tương tác của bạn với người mà bạn quan tâm.',
+    'dashboard_blank_cta' => 'Add your first contact',
+    'dashboard_blank_illustration' => 'Minh họa bởi <a href=":url">Freepik</a>',
+
+    'notes_title' => 'Bạn chưa có ghi chú gắn dấu sao nào.',
+
+    'tab_recent_calls' => 'Cuộc gọi gần đây',
+    'tab_favorite_notes' => 'Ghi chú ưa thích',
+    'tab_calls_blank' => 'Bạn chưa ghi nhật ký cuộc gọi nào.',
+    'tab_debts' => 'Khoản nợ',
+    'tab_debts_blank' => 'Bạn chưa ghi khoản nợ nào.',
+    'tab_tasks' => 'Tasks',
+    'tab_tasks_blank' => 'You haven’t any task yet.',
+
+    'tasks_add_task_placeholder' => 'What is this task about?',
+    'tasks_tab_your_contacts' => 'Tasks related to your contacts',
+    'tasks_tab_your_tasks' => 'Your tasks',
+    'tasks_add_note' => 'Press <kbd>Enter</kbd> to add the task.',
+    'task_add_cta' => 'Add a task',
+
+    'debts_you_owe' => 'You owe',
+
+    'statistics_contacts' => 'Contacts',
+    'statistics_activities' => 'Hoạt động',
+    'statistics_gifts' => 'Quà tặng',
+
+    'reminders_next_months' => 'Sự kiện trong 3 tháng tới',
+    'reminders_none' => 'Không có nhắc nhở nào trong tháng này',
+
+    'product_changes' => 'Cập nhật sản phẩm',
+    'product_view_details' => 'Xem chi tiết',
+];