+++ tests/Commands/PassportCommandTest.php
@@ -3,7 +3,8 @@
 namespace Tests\Commands;
 
 use Tests\TestCase;
-use App\Console\Commands\Passport;
+use Illuminate\Support\Facades\Artisan;
+use App\Console\Commands\Helpers\Command;
 use Laravel\Passport\PersonalAccessClient;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
@@ -14,54 +15,48 @@ class PassportCommandTest extends TestCase
     /** @test */
     public function passport_command_create()
     {
+        /** @var \Tests\Helpers\CommandCallerFake */
+        $fake = Command::fake();
+
         $app = $this->createApplication();
         $app->make('config')->set(['passport.private_key' => '', 'passport.public_key' => '']);
         foreach (PersonalAccessClient::all() as $client) {
             $client->delete();
         }
 
-        $commandExecutor = new CommandExecutorTester();
-        $command = new Passport();
-        $command->commandExecutor = $commandExecutor;
-        $command->setLaravel($app);
+        Artisan::call('monica:passport');
 
-        $command->run(new \Symfony\Component\Console\Input\ArrayInput([]), new \Symfony\Component\Console\Output\NullOutput());
-
-        $this->assertCount(1, $commandExecutor->buffer);
-        $this->assertCommandContains($commandExecutor->buffer[0], '✓ Creating personal access client', 'php artisan passport:client');
+        $this->assertCount(1, $fake->buffer);
+        $this->assertCommandContains($fake->buffer[0], '✓ Creating personal access client', 'php artisan passport:client');
     }
 
     /** @test */
     public function passport_command_already_created()
     {
+        /** @var \Tests\Helpers\CommandCallerFake */
+        $fake = Command::fake();
+
         $app = $this->createApplication();
         $app->make('config')->set(['passport.private_key' => '', 'passport.public_key' => '']);
         PersonalAccessClient::create();
 
-        $commandExecutor = new CommandExecutorTester();
-        $command = new Passport();
-        $command->commandExecutor = $commandExecutor;
-        $command->setLaravel($app);
+        Artisan::call('monica:passport');
 
-        $command->run(new \Symfony\Component\Console\Input\ArrayInput([]), new \Symfony\Component\Console\Output\NullOutput());
-
-        $this->assertCount(0, $commandExecutor->buffer);
+        $this->assertCount(0, $fake->buffer);
     }
 
     /** @test */
     public function passport_command_env_config()
     {
+        /** @var \Tests\Helpers\CommandCallerFake */
+        $fake = Command::fake();
+
         $app = $this->createApplication();
         $app->make('config')->set(['passport.private_key' => '-', 'passport.public_key' => '-']);
 
-        $commandExecutor = new CommandExecutorTester();
-        $command = new Passport();
-        $command->commandExecutor = $commandExecutor;
-        $command->setLaravel($app);
+        Artisan::call('monica:passport');
 
-        $command->run(new \Symfony\Component\Console\Input\ArrayInput([]), new \Symfony\Component\Console\Output\NullOutput());
-
-        $this->assertCount(0, $commandExecutor->buffer);
+        $this->assertCount(0, $fake->buffer);
     }
 
     private function assertCommandContains($array, $message, $command)