+++ CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ### New features:
 
-*
+* Add the notion of AddressBooks
+* Allow customization of life event types
 
 ### Enhancements:
 
-*
+* Use Unknown as default gender
 
 ### Fixes:
 
-*
+* Fix update activity with emotions
+* Fix bad return with wrong credential on oauth/login api
 
 
 # RELEASED VERSIONS: